requirements.txt
numpy==1.22.3
pandas==1.4.3
Pillow==7.1.2
scikit_learn==1.1.1
tqdm==4.64.0
transformers==4.21.1
